requirements.txt
# Core dependencies
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# PDF processing
PyMuPDF>=1.23.0
pdf2image>=1.16.0
Pillow>=10.0.0

# Data processing
pandas>=2.1.0
pyarrow>=14.0.0
numpy>=1.24.0

# API clients
httpx>=0.25.0
aiofiles>=23.0.0
tenacity>=8.2.0

# Validation
jsonschema>=4.20.0

# Image processing
opencv-python>=4.8.0
scikit-image>=0.22.0

# Utilities
tqdm>=4.66.0
click>=8.1.0
PyYAML>=6.0.0
rich>=13.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development
black>=23.0.0
flake8>=6.1.0
mypy>=1.7.0
pre-commit>=3.5.0